+++ app.py
@@ -1,15 +1,29 @@
-from datetime import time, datetime
-import json
 from flask import Flask
 from flask_graphql import GraphQLView
+from flask_graphql import GraphQLView
 from graphene import Schema
-import src.data
-from src.app import app
-from src.schema import Query
+from src.database import db_session, init_db
+from src.schema import schema, Query
+
+app = Flask(__name__)
+app.debug = True
 
 schema = Schema(query=Query)
 
-app.add_url_rule("/", view_func=GraphQLView.as_view("graphql", schema=schema, graphiql=True))
+app.add_url_rule("/graphql", view_func=GraphQLView.as_view("graphql", schema=schema, graphiql=True))
+
+@app.teardown_appcontext
+def shutdown_session(exception=None):
+    db_session.remove()
+
+# Init gym tables and populate them with basic gym information
+init_db()
+db_session.commit()
+
+# scrape C2C website for latest capacity
+from src.c2c_scraper import scrape_capacity
+scrape_capacity()
+db_session.commit()
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=5000)
