--- manage.py
@@ -1,12 +1,0 @@
-import os
-from flask_script import Manager
-from flask_migrate import Migrate, MigrateCommand
-from src.app import app, db
-
-# Build manager
-migrate = Migrate(app, db)
-manager = Manager(app)
-manager.add_command("db", MigrateCommand)
-
-if __name__ == "__main__":
-    manager.run()
