+++ requirements.txt
@@ -1,13 +1,21 @@
 alembic==1.0.9
 aniso8601==3.0.2
 -e git+https://github.com/cuappdev/appdev.py.git@fffe58a01ba5bc00493213e2baa00a6a9e393280#egg=appdev.py
+appdirs==1.4.4
+aspy.yaml==1.3.0
+attrs==23.1.0
 beautifulsoup4==4.6.3
 black==19.3b0
 bs4==0.0.1
 cachetools==3.1.0
 certifi==2018.8.24
+cfgv==3.3.1
 chardet==3.0.4
+charset-normalizer==3.2.0
 click==6.7
+distlib==0.3.7
+fastdiff==0.3.0
+filelock==3.12.2
 Flask==1.0.2
 Flask-GraphQL==2.0.0
 Flask-Migrate==2.4.0
@@ -15,11 +23,14 @@ Flask-Script==2.0.5
 Flask-SQLAlchemy==2.3.1
 google-auth==1.6.3
 graphene==2.1.3
+graphene-sqlalchemy==2.3.0
 graphql-core==2.1
 graphql-relay==0.4.5
 graphql-server-core==1.1.1
 gunicorn==19.9.0
+identify==2.5.24
 idna==2.6
+importlib-metadata==6.7.0
 itsdangerous==0.24
 jedi==0.18.2
 Jinja2==2.10.1
@@ -28,22 +39,35 @@ Mako==1.0.9
 MarkupSafe==1.1.1
 marshmallow==3.0.0rc4
 marshmallow-sqlalchemy==0.16.2
+nodeenv==1.8.0
+parso==0.8.3
+platformdirs==3.10.0
 pre-commit==1.18.3
-promise==2.1
+promise==2.3
 prompt-toolkit==3.0.36
 psycopg2==2.9.5
 ptyprocess==0.7.0
 pure-eval==0.2.2
+pyasn1==0.5.0
+pyasn1-modules==0.3.0
 python-dateutil==2.8.0
 python-editor==1.0.4
 PyYAML==6.0
 requests==2.28.2
 rsa==4.0
 Rx==1.6.1
+singledispatch==3.7.0
 six==1.11.0
-SQLAlchemy==1.3.0
 snapshottest==0.6.0
+SQLAlchemy==1.3.0
 termcolor==2.2.0
+toml==0.10.2
 typing==3.6.6
+typing-extensions==4.7.1
 urllib3==1.26.14
+virtualenv==20.24.3
+wasmer==1.1.0
+wasmer-compiler-cranelift==1.1.0
+wcwidth==0.2.6
 Werkzeug==0.15.5
+zipp==3.15.0
