+++ src/schema.py
@@ -1,189 +1,133 @@
-import datetime as dt
+import datetime as dt 
 
+import graphene
 from graphene import Field, ObjectType, String, List, Int, Boolean
+from graphene import relay
 from graphene.types.datetime import Date, Time
+from graphene_sqlalchemy import SQLAlchemyConnectionField, SQLAlchemyObjectType
+from sqlalchemy import desc
+from src.models.gym import Gym as GymModel, GymTime as GymTimeModel
+from src.models.daytime import DayTime as DayTimeModel
+from src.models.activity import Activity as ActivityModel, Price as PriceModel
+from src.models.capacity import Capacity as CapacityModel
+from src.models.activity import Amenity as AmenityModel
 
-from src.utils import generate_id
 
+class Gym(SQLAlchemyObjectType):
+    class Meta:
+        model = GymModel
 
-class Data(object):
-    gyms = {}
-    classes_by_date = {}
-    class_details = {}
-
-    @staticmethod
-    def update_data(**kwargs):
-        Data.gyms = kwargs.get("gyms")
-
-        date_limit = dt.date.today() - dt.timedelta(days=kwargs.get("limit"))
-        classes_by_date = Data.classes_by_date
-        classes_data = kwargs.get("classes")
-        for class_id, class_data in classes_data.items():
-            if class_data.date in classes_by_date:
-                classes_by_date[class_data.date][class_id] = class_data
-            else:
-                classes_by_date[class_data.date] = {class_id: class_data}
-        classes_by_date = {date: classes_data for date, classes_data in classes_by_date.items() if date > date_limit}
-        Data.classes_by_date = classes_by_date
-
-        Data.class_details = kwargs.get("class_details")
+    times = graphene.List(lambda: DayTime, day=graphene.Int(), start_time=graphene.DateTime(), end_time=graphene.DateTime(), restrictions=graphene.String(), special_hours=graphene.Boolean())
+    activities = graphene.List(lambda: Activity, name=graphene.String())
+    capacities = graphene.List(lambda: Capacity, gym_id=graphene.Int())
 
     @staticmethod
-    def update_pool_hours(gyms, pool_hours):
-        for gym in [gyms[generate_id("Helen Newman")], gyms[generate_id("Teagle Down")]]:
-            facility = FacilityType(details=[], name="Pool")
-            if gym.name in pool_hours:
-                days = pool_hours[gym.name]
-                created = False
-                for idx, day in enumerate(days, 1):
-                    for hours in day:
-                        d = idx % 7
-                        if created:
-                            details = facility.details[0]
-                            time = next((time for time in details.times if time.day == d), None)
-                            if time:
-                                time.time_ranges.append(hours)
-                            else:
-                                details.times.append(DayTimeRangesType(day=d, time_ranges=[hours]))
-                        else:
-                            created = True
-                            facility.details.append(
-                                FacilityDetailsType(
-                                    details_type="Hours",
-                                    equipment=[],
-                                    image_urls=[],
-                                    items=[],
-                                    prices=[],
-                                    sub_facility_names=[],
-                                    times=[DayTimeRangesType(day=d, time_ranges=[hours])],
-                                )
-                            )
-                            gym.facilities.append(facility)
+    def resolve_times(self, info, day=None, start_time=None, end_time=None):
+        query = GymTime.get_query(info=info)
+        query = query.filter(GymTimeModel.gym_id == self.id)
+        query_daytime = DayTime.get_query(info=info) #could be wrong
+        if day:
+          query_daytime = query_daytime.filter(DayTimeModel.day == day)
+        if start_time:
+          query_daytime = query_daytime.filter(DayTimeModel.start_time == start_time)
+        if end_time:
+          query_daytime = query_daytime.filter(DayTimeModel.end_time == end_time)
+        
+        daytime_queries = []
+        for row in query:
+          daytime = query_daytime.filter(DayTimeModel.id == row.daytime_id)
+          if daytime.first():
+            daytime_queries.append(daytime[0])
 
+        return daytime_queries
 
-class DayTimeRangeType(ObjectType):
-    day = Int(required=True)
-    end_time = Time(required=True)
-    restrictions = String(default_value="", required=True)
-    special_hours = Boolean(default_value=False, required=True)
-    start_time = Time(required=True)
+    def resolve_activities(self, info, name=None):
+        query = Activity.get_query(info=info)
+        activity_queries = []
+        for act in self.activities:
+            activity = query.filter(ActivityModel.id == act.id)
+            if activity.first() and (name == act.name or name == None):
+                    activity_queries.append(activity[0])
+        return activity_queries
+    @staticmethod
+    def resolve_capacities(self, info, gym_id = None):
+      query = Capacity.get_query(info=info) \
+        .filter(CapacityModel.gym_id == self.id) \
+        .order_by(desc(CapacityModel.updated))
 
+      return [query.first()]
+        
+class DayTime(SQLAlchemyObjectType):
+  class Meta:
+    model = DayTimeModel
 
-class TimeRangeType(ObjectType):
-    end_time = Time(required=True)
-    restrictions = String(default_value="", required=True)
-    special_hours = Boolean(default_value=False, required=True)
-    start_time = Time(required=True)
+class GymTime(SQLAlchemyObjectType):
+  class Meta:
+    model = GymTimeModel
 
+class Activity(SQLAlchemyObjectType):
+    class Meta:
+        model = ActivityModel
 
-class DayTimeRangesType(ObjectType):
-    day = Int(required=True)
-    time_ranges = List(TimeRangeType, required=True)
+    gyms = graphene.List(lambda: Gym, name=graphene.String())
+    prices = graphene.List(lambda: Price, cost=graphene.Int(), one_time=graphene.Boolean())
 
+    def resolve_gyms(self, info, name=None):
+        query = Gym.get_query(info=info)
+        gym_queries = []
+        for g in self.gyms:
+            gym = query.filter(GymModel.id == g.id)
+            if gym.first() and (name == g.name or name == None):
+                gym_queries.append(gym[0])
+              
+        return gym_queries
 
-class EquipmentType(ObjectType):
-    equipment_type = String(required=True)
-    name = String(required=True)
-    quantity = String(default_value="1", required=True)
-    workout_type = String(required=True)
+class Capacity(SQLAlchemyObjectType):
+  class Meta:
+    model = CapacityModel
+  
 
+  def resolve_prices(self, info, cost=None, one_time=None):
+    query = Price.get_query(info=info)
+    query = query.filter(PriceModel.activity_id == self.id)
+    if cost:
+      query = query.filter(PriceModel.cost == cost)
+    if one_time is not None:
+      query = query.filter(PriceModel.one_time == one_time)
+    return query
 
-# details types: Equipment, Hours, Images, Phone Numbers, Prices, Sub-Facilities
-class FacilityDetailsType(ObjectType):
-    details_type = String(required=True)
-    equipment = List(EquipmentType, required=True)
-    image_urls = List(String, required=True)
-    items = List(String, required=True)
-    prices = List(String, required=True)
-    sub_facility_names = List(String, required=True)
-    times = List(DayTimeRangesType, required=True)
+class Price(SQLAlchemyObjectType):
+  class Meta:
+    model = PriceModel
 
+class Amenity(SQLAlchemyObjectType):
+  class Meta:
+    model = AmenityModel
 
-class FacilityType(ObjectType):
-    details = List(FacilityDetailsType, required=True)
-    name = String(required=True)
+class Query(graphene.ObjectType):
 
+    gyms = graphene.List(lambda: Gym, 
+      id=graphene.Int(),
+      name=graphene.String(), 
+      description=graphene.String(), 
+      location=graphene.String(),
+      latitude=graphene.Float(),
+      longitude=graphene.Float(),
+      image_url=graphene.String())
 
-class GymType(ObjectType):
-    id = String(required=True)
-    name = String(required=True)
-    description = String(required=True)
-    facilities = List(FacilityType, required=True)
-    popular = List(List(Int))
-    times = List(DayTimeRangeType, required=True)
-    image_url = String()
+    def resolve_gyms(self, info, name=None):
+      query = Gym.get_query(info)
+      if name:
+        query=query.filter(GymModel.name == name)
+      return query.all()
 
-    def is_open(self, day=None):
-        return day is None or any([day == dt_range.day for dt_range in self.times])
+    activities = graphene.List(lambda: Activity, name=graphene.String( \
+      ), details=graphene.String(), image_url=graphene.String())
 
+    def resolve_activities(self, info, name=None):
+        query = Activity.get_query(info)
+        if name:
+            query = query.filter(ActivityModel.name == name)
+        return query.all()
 
-class TagType(ObjectType):
-    label = String(required=True)
-    image_url = String(required=True)
-
-
-class ClassDetailType(ObjectType):
-    id = String(required=True)
-    name = String(required=True)
-    description = String(required=True)
-    tags = List(TagType, required=True)
-    categories = List(String, required=True)
-
-
-class ClassType(ObjectType):
-    id = String(required=True)
-    gym_id = String()
-    gym = Field(GymType)
-    location = String(required=True)
-    details_id = String(required=True)
-    details = Field(ClassDetailType, required=True)
-    date = Date(required=True)
-    start_time = Time()
-    end_time = Time()
-    instructor = String(required=True)
-    is_cancelled = Boolean(required=True)
-    image_url = String(required=True)
-
-    def resolve_gym(self, info):
-        return Data.gyms.get(self.gym_id)
-
-    def resolve_details(self, info):
-        return Data.class_details.get(self.details_id)
-
-    def filter(self, detail_ids=None, day=None, name=None, tags=None, gym_id=None, instructor=None):
-        details = Data.class_details.get(self.details_id)
-        return (
-            (detail_ids is None or self.details_id in detail_ids)
-            and (day is None or day == self.date)
-            and (name is None or name in details.name)
-            and (tags is None or any([tag in details.tags for tag in tags]))
-            and (gym_id is None or gym_id == self.gym_id)
-            and (instructor is None or instructor in self.instructor)
-        )
-
-
-class Query(ObjectType):
-    gyms = List(GymType, day=Date(), gym_id=String(name="id"))
-    classes = List(
-        ClassType,
-        detail_ids=List(String),
-        day=Date(),
-        name=String(),
-        tags=List(String),
-        gym_id=String(),
-        instructor=String(),
-    )
-
-    def resolve_gyms(self, info, day=None, gym_id=None):
-        if gym_id is not None:
-            gym = Data.gyms.get(gym_id)
-            return [gym] if gym is not None else []
-        return [gym for gym in Data.gyms.values() if gym.is_open(day)]
-
-    def resolve_classes(self, info, **kwargs):
-        result = []
-        for classes in Data.classes_by_date.values():
-            for c in classes.values():
-                if c.filter(**kwargs):
-                    result.append(c)
-        return result
+schema = graphene.Schema(query=Query)
